Fix an issue with reused producers erroring.
Previously the returned messages queue was used upon RPC delivery
to check if the message arrived at the destination or not.  As
this was never cleared however, a reused producer on a different
queue will result in an incorrectly delivered error message.

Without this delivering a message to a missing queue will cause
the service to malfunction even on working queues.