prevent unsupported recursive env var lookups
switching to safe loading changes the behaviour of the default regex
parser when you give it a config that contains (unsupported) recursive
environment variables.

rather than update the tests to assert a different broken behaviour,
we now raise a configuration error. this will alert the user that their
config file will likely not be parsed correctly, or they're missing
the optional library.