check_for_lost_replies needs a retry policy
https://github.com/nameko/nameko/pull/718 fixes the standalone rpc
reply listener to honour the amqp uri passed into the rpc client.
the reconnection tests have revelealed another bug in the
implementation now that the reply listener is using the correct uri.

check_for_lost_replies will see a "server unexpectedly closed" OSError
when reconnecting after a broken connection. this is because it uses
an existing connection from the connection pool, some of which will
have previously been connected.

the above PR was merged without a green CI build (because _reasons_),
so the fix is being addded as a separate commit.